requirements.txt
google-adk>=1.0.0
mcp>=1.9.0
python-dotenv>=1.0.1
